Fixed Short type.

+++ modules/cassandra/store/src/test/java/org/apache/ignite/tests/utils/TestsHelper.java
@@ -586,7 +586,8 @@ public class TestsHelper {
         for (int i = 0; i < phonesCnt; i++)
             phones.add(randomNumber(4));
 
-        return new Person(personNum, randomString(4), Long.toString(RANDOM.nextInt(100)), RANDOM.nextInt(100),
+        return new Person(personNum, randomString(4), Long.toString(RANDOM.nextInt(100)),
+                (short)RANDOM.nextInt(100),
             RANDOM.nextBoolean(), RANDOM.nextLong(), RANDOM.nextFloat(), new Date(), phones);
     }
